clean up pid/workspace stuff
